스터디룸 반응형 구현 및 기능 구현
