+++ src/components/common/PageHeader.jsx
@@ -3,25 +3,21 @@ import styled from 'styled-components';
 import backGroundUrl from '../../assets/images/mypage/mypageBackground.png';
 import { PuppleButton } from '../style/Button';
 
-// PageHeader 컴포넌트 정의, 이 컴포넌트는 페이지 헤더를 렌더링합니다.
-const PageHeader = ({ pageTitle, headerTitles, activeButtonIndex, onButtonClick, changeColorOnClick, changeColorOnHover }) => {
+const PageHeader = ({ pageTitle, headerTitles, activeButtonIndex, onButtonClick, changeColorOnClick, changeColorOnHover, large }) => {
     return (
         <>
-            {/* Header 스타일을 가진 div로 헤더 영역을 감쌉니다. */}
             <Header pageTitle={pageTitle}>
-                {/* 페이지 제목을 표시합니다. */}
                 <PageHeaderTitle>{pageTitle}</PageHeaderTitle>
-                <RowWrapper>
-                    {/* headerTitles 배열을 순회하며 버튼을 생성합니다. */}
+                <RowWrapper large={large}>
                     {headerTitles.map((title, index) => (
                         <StyledPuppleButton
                             key={index}
-                            $isActive={activeButtonIndex === index} // 현재 활성화된 버튼인지 확인합니다.
-                            $changeColorOnClick={changeColorOnClick} // 클릭 시 색상을 변경할지 여부를 결정합니다.
-                            $changeColorOnHover={changeColorOnHover} // 호버 시 색상을 변경할지 여부를 결정합니다.
-                            onClick={() => onButtonClick(index)} // 버튼 클릭 시 실행할 함수입니다.
+                            $isActive={activeButtonIndex === index}
+                            $changeColorOnClick={changeColorOnClick}
+                            $changeColorOnHover={changeColorOnHover}
+                            onClick={() => onButtonClick(index)}
                         >
-                            {title} {/* 버튼의 텍스트입니다. */}
+                            {title}
                         </StyledPuppleButton>
                     ))}
                 </RowWrapper>
@@ -32,13 +28,11 @@ const PageHeader = ({ pageTitle, headerTitles, activeButtonIndex, onButtonClick,
 
 export default PageHeader;
 
-// Header 스타일을 정의합니다. 배경 이미지와 색상을 포함합니다.
 const Header = styled.div`
-    display: flex;;
-    z-index: 2;
-    position: ${({ pageTitle }) => ( pageTitle === '마이페이지' ? 'fixed' : 'relative')};
-    top : ${({ pageTitle }) => ( pageTitle === '마이페이지' ? '60px' : '')};
     display: flex;
+    z-index: 2;
+    position: ${({ pageTitle }) => (pageTitle === '마이페이지' ? 'fixed' : 'relative')};
+    top: ${({ pageTitle }) => (pageTitle === '마이페이지' ? '60px' : '')};
     flex-direction: column;
     align-items: center;
     justify-content: center;
@@ -46,15 +40,11 @@ const Header = styled.div`
     height: 16.1875em;
     background-color: #FBFAFF;
     background-image: url(${backGroundUrl});
-
     @media (max-width: 768px) {
-        top: 80px;
-        height: auto;
-        padding: 2em 1em;
+        margin-bottom : ${({ large }) => (large ? '0' : '3em')};
     }
 `;
 
-// PageHeaderTitle 스타일을 정의합니다. 제목의 폰트 크기와 색상을 포함합니다.
 const PageHeaderTitle = styled.div`
     margin-top: 1.0625em;
     margin-bottom: 1.5625em;
@@ -69,14 +59,22 @@ const PageHeaderTitle = styled.div`
     }
 `;
 
-// RowWrapper 스타일을 정의합니다. 버튼들을 가로로 나열합니다.
 const RowWrapper = styled.div`
     display: flex;
-    justify-content: center;
     gap: 1em;
+    justify-content: center;
+
+    @media (max-width: 768px) {
+        display: ${({ large }) => (large ? 'grid' : 'flex')};
+        grid-template-columns: ${({ large }) => (large ? '1fr 1fr' : '')};
+        place-items : center center;
+        width : auto;
+        
+        font-size: 1em;
+    }
+
 `;
 
-// StyledPuppleButton 스타일을 정의합니다. 버튼의 크기, 색상, 폰트 스타일을 포함합니다.
 const StyledPuppleButton = styled(PuppleButton)`
     box-sizing: border-box;
     width: 10em;
@@ -91,14 +89,8 @@ const StyledPuppleButton = styled(PuppleButton)`
     }
 
     @media (max-width: 768px) {
-        width: 5em;
+        width : ${({ large }) => (large ? '3em' : '10em')};
         height: 3em;
-        font-size: 1em;
-    }
-    @media (min-width: 768px) and (max-width: 991px) {
-        width: 10em;
-        height: 3em;
-        font-size: 1em;
+        font-size: 0.8125em;
     }
 `;
-
